Fix bug, description:
When grid was re-added to the root, it lost its matrix data.
Even when the buffer loads old matrix data upon geometry recording,
the data is lost, because the buffer is recorded one times without the grid spatial. Therefore there is no old data to load.
Fix the problem by introducing visibility, which makes a geometry invisible, like a spatial. This requires geometry buffer re-recording, but keeps the data within the buffer.