0.0.0-unstable.21 Implement smoothed difference delegation
If used, property read will return the value difference since the last read operation, instead of the current absolute value. To keep track of the absolute value, you have to accumulate all read retrievals, which implies that you have to define explicitly who can read the property at all. Otherwise, the accumulated absolute value will get corrupt.

This method of smoothing is intended to be used when dropping pre-defined transform matrices. Since the absolute rotation is no longer stored as state in a geometry, since that transform has already been merged into the older model matrices, we want to rotate by the difference, not to an absolute value.