0.0.0-relax-fixed-transform-matrices.22+8 Finalize this feature branch
Refactor Geometry

Currently, transforming is not smoothed.
Geometries can no longer do the out-of-box, since rotation is no longer stored absolutely, but set in small sub-steps. As soon as the next rotate() call occur, the previous rotation value has vanished since rotation matrices are merged.
So one way to implement this would be to move the smoothed properties to the controllers, but that would require to constantly query them. That will be fixed in a future feature branch.