Reimplement attraction gesture listener in a far more genetic form
Previously, the attraction detector implemented a specific interface, which restricts applicability of the scroll attractor to a specific gesture detector, as they all call different on... listener-functions.
Now, the attraction class is completely independent, the user only has to call manually scrollTo() and resetScrollTo() in order to manipulator the attractors position.